Python: Port xml.sax tests
